src/requirements.txt
pandas==2.1.1
numpy==1.26.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.0
